[chore] Components reorganized
