--- src/routes/BillingProfiles.js
@@ -1,15 +1,0 @@
-import React, { Component, Fragment } from "react";
-import BillingProfileList from "../components/Sales/BillingProfileList";
-
-class BillingProfiles extends Component {
-
-    render() {
-        return (
-            <Fragment>
-                <BillingProfileList></BillingProfileList>
-            </Fragment>
-        );
-    }
-}
-
-export default BillingProfiles;
